Cleanup after merge; moved files to packages
